feat search tokens (#60)
* feat(domain): implement create search token

Signed-off-by: luminuszz <luminuszz43@gmail.com>

* feat(infra): create database schema and command and controller handler for create search token

Signed-off-by: luminuszz <luminuszz43@gmail.com>

* feat: add createMany method to SearchTokenRepository

This commit adds a new method, createMany, to the SearchTokenRepository abstract class. The createMany method allows for the creation of multiple search tokens at once by accepting an array of SearchToken objects. This change was made to improve efficiency and provide a more convenient way to create multiple search tokens in a single operation.

Signed-off-by: luminuszz <luminuszz43@gmail.com>

* feat: list search tokens by type

Signed-off-by: luminuszz <luminuszz43@gmail.com>

* chore: add swagger definition for SearchToken

Signed-off-by: luminuszz <luminuszz43@gmail.com>

---------

Signed-off-by: luminuszz <luminuszz43@gmail.com>